+++ CHANGELOG.md
@@ -5,10 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [0.0.1] - 2018-05-30
+## [0.2.0] - 2018-05-25
 ### Added
-- Initial commit
+- Build keyring
+- Add package extras
 
 ## [0.1.0] - 2018-05-24
 ### Added
 - Beta release
+
+## [0.0.1] - 2018-05-30
+### Added
+- Initial commit